Create separate hook for search creation.
